pyproject.toml
[tool.poetry]
name = "drf-boilerplate"
version = "0.1.0"
description = ""
authors = ["Michael Youngchan CHO <michael.yc.cho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
flex = "^6.14.1"
setuptools = "^70.3.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
